refactored gridview dialog into it's own class
