+++ app/src/main/res/values/strings.xml
@@ -62,4 +62,5 @@
     <string name="edit_fragment_image">Image</string>
     <string name="edit_fragment_image_title">Image Editor</string>
     <string name="fragment_edit_control_switch_button">Button Type</string>
+    <string name="image_grid_title">Choose Image</string>
 </resources>
